+++ modules/agar/src/test/ts/module/ephox/agar/test/DomContainers.ts
@@ -1,7 +1,7 @@
-import { Attribute, Insert, Remove, SugarElement } from '@ephox/sugar';
+import { Attribute, Insert, Remove, SugarBody, SugarElement } from '@ephox/sugar';
 import { Step } from 'ephox/agar/api/Step';
 
-const mSetup = Step.stateful((state, next, _die) => {
+const createContainer = () => {
   const container = SugarElement.fromTag('div');
   Attribute.set(container, 'tabindex', '-1');
   Attribute.set(container, 'test-id', 'true');
@@ -9,10 +9,33 @@ const mSetup = Step.stateful((state, next, _die) => {
   const input = SugarElement.fromTag('input');
   Insert.append(container, input);
 
-  Insert.append(SugarElement.fromDom(document.body), container);
+  return {
+    container,
+    input
+  };
+};
+
+const mSetup = Step.stateful((state, next, _die) => {
+  const { container, input } = createContainer();
+  Insert.append(SugarBody.body(), container);
   next({
     container,
     input
+  });
+});
+
+const mSetupShadowRoot = Step.stateful((state, next, _die) => {
+  const { container, input } = createContainer();
+
+  const shadowHost = SugarElement.fromTag('div');
+  const shadowRoot = SugarElement.fromDom(shadowHost.dom.attachShadow({ mode: 'open' }));
+  Insert.append(shadowRoot, container);
+
+  Insert.append(SugarBody.body(), shadowHost);
+  next({
+    container,
+    input,
+    shadowRoot
   });
 });
 
@@ -27,5 +50,6 @@ const mTeardown = Step.stateful<TeardownState, TeardownState>((state, next, _die
 
 export {
   mSetup,
+  mSetupShadowRoot,
   mTeardown
 };