+++ modules/alloy/src/main/ts/ephox/alloy/aria/AriaFocus.ts
@@ -1,10 +1,10 @@
 import { Fun, Optional } from '@ephox/katamari';
-import { Compare, Focus, PredicateFind, SugarElement, Traverse } from '@ephox/sugar';
+import { Compare, Focus, PredicateFind, SugarElement, SugarShadowDom } from '@ephox/sugar';
 
 const preserve = <T>(f: (e: SugarElement) => T, container: SugarElement): T => {
-  const ownerDoc = Traverse.owner(container);
+  const dos = SugarShadowDom.getRootNode(container);
 
-  const refocus = Focus.active(ownerDoc).bind((focused: SugarElement) => {
+  const refocus = Focus.active(dos).bind((focused: SugarElement) => {
     const hasFocus = (elem: SugarElement) => Compare.eq(focused, elem);
     return hasFocus(container) ? Optional.some(container) : PredicateFind.descendant(container, hasFocus);
   });
@@ -13,7 +13,7 @@ const preserve = <T>(f: (e: SugarElement) => T, container: SugarElement): T => {
 
   // If there is a focussed element, the F function may cause focus to be lost (such as by hiding elements). Restore it afterwards.
   refocus.each((oldFocus: SugarElement) => {
-    Focus.active(ownerDoc).filter((newFocus) => Compare.eq(newFocus, oldFocus)).fold(() => {
+    Focus.active(dos).filter((newFocus) => Compare.eq(newFocus, oldFocus)).fold(() => {
       // Only refocus if the focus has changed, otherwise we break IE
       Focus.focus(oldFocus);
     }, Fun.noop);